remove excess containers
