+++ frontend/features/jams/components/jammer-state-editor.tsx
@@ -7,7 +7,10 @@ import {
   MantineProvider,
 } from "@mantine/core";
 import { IconStarFilled } from "@tabler/icons-react";
-import { UseMutationResult } from "@tanstack/react-query";
+import { useTeamJamAddLead } from "../hooks/use-team-jam-add-lead";
+import { useTeamJamAddLost } from "../hooks/use-team-jam-add-lost";
+import { useTeamJamAddStarPass } from "../hooks/use-team-jam-add-star-pass";
+import { TeamJam } from "@/types/jam";
 
 const checkBoxTheme = createTheme({
   // Hovering over checkbox should change cursor
@@ -32,18 +35,12 @@ interface JammerStateProps extends GroupProps {
    * the other team's Jammer has been declared lead.
    */
   isLeadEligible: boolean;
-  /**
-   * The event handler which fires when clicking the lead checkbox.
-   */
-  leadOnClick?: UseMutationResult<void, unknown, boolean, unknown>;
-  /**
-   * The event handler which fires when clicking the lost checkbox.
-   */
-  lostOnClick?: UseMutationResult<void, unknown, boolean, unknown>;
-  /**
-   * The event handler which fires when clicking the star pass checkbox.
-   */
-  starPassOnClick?: UseMutationResult<void, unknown, boolean, unknown>;
+
+  boutUuid: string;
+  periodNum: number;
+  jamNum: number;
+  teamNum: number;
+  teamJam: TeamJam;
 }
 
 /**
@@ -52,15 +49,37 @@ interface JammerStateProps extends GroupProps {
  * occurred.
  */
 export default function JammerStateEditor({
-  lead,
-  lost,
-  starPass,
   isLeadEligible,
-  leadOnClick,
-  lostOnClick,
-  starPassOnClick,
+  boutUuid,
+  periodNum,
+  jamNum,
+  teamNum,
+  teamJam,
   ...props
 }: JammerStateProps) {
+  const setLead = useTeamJamAddLead({
+    boutUuid,
+    periodNum,
+    jamNum,
+    teamNum,
+  });
+  const setLost = useTeamJamAddLost({
+    boutUuid,
+    periodNum,
+    jamNum,
+    teamNum,
+  });
+  const setStarPass = useTeamJamAddStarPass({
+    boutUuid,
+    periodNum,
+    jamNum,
+    teamNum,
+  });
+
+  const lead = teamJam.events.some((tripEvent) => tripEvent.lead);
+  const lost = teamJam.events.some((tripEvent) => tripEvent.lost);
+  const starPass = teamJam.events.some((tripEvent) => tripEvent.starPass);
+
   return (
     <MantineProvider theme={checkBoxTheme}>
       <Group {...props}>
@@ -68,7 +87,7 @@ export default function JammerStateEditor({
           label="Lead"
           checked={lead}
           disabled={!isLeadEligible}
-          onClick={() => leadOnClick?.mutate(!lead)}
+          onClick={() => setLead.mutate(!lead)}
           variant="outline"
           icon={({ ...others }) => <IconStarFilled {...others} />}
         />
@@ -76,14 +95,14 @@ export default function JammerStateEditor({
         <Checkbox
           label="Lost"
           checked={lost}
-          onClick={() => lostOnClick?.mutate(!lost)}
+          onClick={() => setLost.mutate(!lost)}
           variant="outline"
         />
         <Divider orientation="vertical" />
         <Checkbox
           label="Star Pass"
           checked={starPass}
-          onClick={() => starPassOnClick?.mutate(!starPass)}
+          onClick={() => setStarPass.mutate(!starPass)}
           variant="outline"
         />
       </Group>
