+++ frontend/features/jams/components/pass-editor.tsx
@@ -1,5 +1,5 @@
 import { Button, Group, GroupProps } from "@mantine/core";
-import { UseMutationResult } from "@tanstack/react-query";
+import { useTeamJamAddTrip } from "../hooks/use-team-jam-add-trip";
 
 interface PassEditorProps {
   /**
@@ -10,10 +10,10 @@ interface PassEditorProps {
    * The number of passes allowed in a trip.
    */
   numPasses: number;
-  /**
-   * Mutator which handles adding passes to this TeamJam.
-   */
-  addPassOnClick?: UseMutationResult<void, unknown, number, unknown>;
+  boutUuid: string;
+  periodNum: number;
+  jamNum: number;
+  teamNum: number;
 }
 
 /**
@@ -24,16 +24,26 @@ interface PassEditorProps {
 export default function PassEditor({
   showInitial = false,
   numPasses,
-  addPassOnClick,
+  boutUuid,
+  periodNum,
+  jamNum,
+  teamNum,
   ...props
 }: PassEditorProps & GroupProps) {
+  const addTrip = useTeamJamAddTrip({
+    boutUuid,
+    periodNum,
+    jamNum,
+    teamNum,
+  });
+
   if (showInitial) {
     return (
       <Group justify="center" gap="md" {...props}>
-        <Button variant="subtle" onClick={() => addPassOnClick?.mutate(0)}>
+        <Button variant="subtle" onClick={() => addTrip.mutate(0)}>
           No Pass
         </Button>
-        <Button variant="outline" onClick={() => addPassOnClick?.mutate(4)}>
+        <Button variant="outline" onClick={() => addTrip.mutate(4)}>
           Initial
         </Button>
       </Group>
@@ -46,7 +56,7 @@ export default function PassEditor({
         <Button
           key={i}
           variant={i < numPasses ? "subtle" : "outline"}
-          onClick={() => addPassOnClick?.mutate(i)}
+          onClick={() => addTrip.mutate(i)}
         >
           {i}
         </Button>
